+++ clients/go-client/examples/advanced/main.go
@@ -8,8 +8,8 @@ import (
 	"sync"
 	"time"
 
-	"toolplane-go-client/client"
-	clientpb "toolplane-go-client/proto"
+	"github.com/Abhishek-chohan/tool-control-plane/clients/go-client/client"
+	clientpb "github.com/Abhishek-chohan/tool-control-plane/clients/go-client/proto"
 )
 
 func main() {